LayoutTests/inspector/canvas/create-context-webgl2-expected.txt

Test that CanvasManager tracks creation and destruction of WebGL2 canvases.


== Running test suite: Canvas.CreateContextWebGL2
-- Running test case: Canvas.CreateContextWebGL2.NoCanvases
PASS: CanvasManager should have no canvases.

-- Running test case: Canvas.CreateContextWebGL2.Attached
PASS: Canvas context should be WebGL2.
PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextWebGL2.Detached
PASS: Canvas context should be WebGL2.
PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextWebGL2.CSSCanvas
Create CSS canvas from -webkit-canvas(css-canvas).
PASS: Canvas context should be WebGL2.
PASS: Canvas name should equal the identifier passed to -webkit-canvas.
